Kb rio identify key authors (#5)
* Key Researchers

my code to generate the key researchers

* My code to sort out the top authors

This is my code to sort out the top authors of leptospirosis research and plot them in bar charts. I've added comments too.
The file I uploaded last week was in the wrong format. Sorry.

* Project leptospirosis data

data in .csv that I've used for the graphs

* Create placeholder.txt

* Rename Key_Researchers_Rio (1).ipynb to src/Key_Researchers_Rio (1).ipynb

moved file to src folder

* Rename Skills Project to src/Skills Project

* Create placeholder.txt

* Rename articles.leptospirosis.csv to data/articles.leptospirosis.csv

moved files into folder

* moved file into folder

* moved file into folder

* Rebased and updated the notebook

---------

Co-authored-by: Sneha Lakshminarayanan <snehal2242@gmail.com>